9836: Factor out submitterType util to fix circular deps
